+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Optimized package size
 - [#63](https://github.com/estruyf/vscode-demo-time/issues/63): Added completion and hover panel support for the frontmatter of the markdown slides
+- [#69](https://github.com/estruyf/vscode-demo-time/issues/69): Added the functionality to export the slides to a PDF file
 
 ## [0.0.85] - 2025-03-18